refactor(inserviveis): new inserviveis page with lots of features
