+++ src.mainproc/ipc/escolas/getEscolas.ts
@@ -1,6 +1,6 @@
 import { ipcMain } from "electron";
-import { getEscolas } from "../lib/StorageManager";
-import { getWindow } from "../lib/WindowManager";
+import { getEscolas } from "../../lib/StorageManager";
+import { getWindow } from "../../lib/WindowManager";
 
 ipcMain.on("getEscolas", () => {
   const escolas = getEscolas();
